refactor(ipod): extract useIpodScale from useIpodLogic
