+++ src/apps/ipod/hooks/useIpodLogic.ts
@@ -15,6 +15,7 @@ import { useCustomEventListener, useEventListener } from "@/hooks/useEventListen
 import { useLibraryUpdateChecker } from "./useLibraryUpdateChecker";
 import { useIpodActiveLibrary } from "./useIpodActiveLibrary";
 import { useIpodPlayback } from "./useIpodPlayback";
+import { useIpodScale } from "./useIpodScale";
 import {
   useAppleMusicLibrary,
   syncAppleMusicResource,
@@ -4662,58 +4663,8 @@ export function useIpodLogic({
     [playScrollSound, registerActivity, registerActivityRef, menuMode, menuHistory, isFullScreen, showStatus, isCoverFlowOpen, isMusicQuizOpen, isBrickGameOpen, getMenuItemLetter, scheduleFastScrollIdle, setSelectedMenuItem]
   );
 
-  // Scaling
-  const containerRef = useRef<HTMLDivElement>(null);
-  const [scale, setScale] = useState(1);
-  const prevMinimizedRef = useRef(isMinimized);
-
-  useEffect(() => {
-    let timeoutId: number;
-
-    const handleResize = () => {
-      if (!containerRef.current) return;
-
-      requestAnimationFrame(() => {
-        if (!containerRef.current) return;
-        const containerWidth = containerRef.current.clientWidth;
-        const containerHeight = containerRef.current.clientHeight;
-        const baseWidth = 250;
-        const baseHeight = 400;
-        const availableWidth = containerWidth - 50;
-        const availableHeight = containerHeight - 50;
-        const widthScale = availableWidth / baseWidth;
-        const heightScale = availableHeight / baseHeight;
-        const newScale = Math.min(widthScale, heightScale, 2);
-        const finalScale = Math.max(1, newScale);
-
-        setScale((prevScale) => {
-          if (Math.abs(prevScale - finalScale) > 0.01) return finalScale;
-          return prevScale;
-        });
-      });
-    };
-
-    timeoutId = window.setTimeout(handleResize, 10);
-
-    if (prevMinimizedRef.current && !isMinimized) {
-      [50, 100, 200, 300, 500].forEach((delay) => {
-        window.setTimeout(handleResize, delay);
-      });
-    }
-    prevMinimizedRef.current = isMinimized;
-
-    const resizeObserver = new ResizeObserver(() => {
-      clearTimeout(timeoutId);
-      timeoutId = window.setTimeout(handleResize, 10);
-    });
-
-    if (containerRef.current) resizeObserver.observe(containerRef.current);
-
-    return () => {
-      clearTimeout(timeoutId);
-      resizeObserver.disconnect();
-    };
-  }, [isWindowOpen, isMinimized]);
+  // Scaling (extracted to useIpodScale)
+  const { containerRef, scale } = useIpodScale({ isWindowOpen, isMinimized });
 
   // Share and lyrics handlers
   const handleShareSong = useCallback(() => {
